+++ docs/src/main/paradox/release-notes.md
@@ -6,9 +6,10 @@
 
 * Upgraded to the following core dependencies: Spark 2.3.2, GeoTrellis 2.3.0, GeoMesa 2.2.1, JTS 1.16.0.
 * Build `pyrasterframes` binary distribution for pip installation.
+* Added support for rendering RasterFrame types in IPython/Jupyter.
 * Added new tile functions `rf_round`, `rf_abs`, `rf_log`, `rf_log10`, `rf_log2`, `rf_log1p`, `rf_exp`, `rf_exp10`, `rf_exp2`, `rf_expm1`, `rf_resample`.
-* Support python-side Tile User-Defined Type backed by [numpy](https://www.numpy.org/) `ndarray` or `ma.MaskedArray`.
-* Support python-side [Shapely](https://pypi.org/project/Shapely/) geometry User-Defined Type.
+* Support Python-side Tile User-Defined Type backed by [numpy](https://www.numpy.org/) `ndarray` or `ma.MaskedArray`.
+* Support Python-side [Shapely](https://pypi.org/project/Shapely/) geometry User-Defined Type.
 * SQL API support for: `rf_assemble_tile`, `rf_array_to_tile`.
 * Introduced at the source level the concept of a `RasterSource` and `RasterRef`, enabling lazy/delayed read of sub-scene tiles.
 * Added `withKryoSerialization` extension methods on `SparkSession.Builder` and `SparkConf`.
@@ -19,6 +20,8 @@
 * Added `rf_crs` for `ProjectedRasterTile` columns.
 * Added `st_extent` (for `Geometry` types) and `rf_extent` (for `ProjectedRasterTile` and `RasterSource` columns).
 * Added `st_geometry` (for `Extent` types) and `rf_geometry` (for `ProjectedRasterTile` and `RasterSource` columns).
+* _Breaking_: The type `RasterFrame` renamed `RasterFrameLayer` to be reflect its purpose.
+* _Breaking_: All `asRF` methods renamed to `asLayer`.
 * _Breaking_: Root package changed from `org.locationtech.rasterframes` to `org.locationtech.rasterframes`.
 * _Breaking_: Removed `envelope`, in lieu of `st_extent`, `rf_extent` or `st_envelope` 
 * _Breaking_: Renamed `rf_extent_geometry` to `st_geometry`